+++ src/internal/pool/mod.rs
@@ -1,5 +1,6 @@
 mod api;
 mod config;
+mod constructor;
 mod drop_impl;
 mod restart_guard;
 mod shared;